Solver adapted to work with tkinter
Made the backtracking solver work with the GUI so it shows each step as it does it.

Also changed how the numbers that were set at the start of the game so they can't be changed by any means (there was a way to before by selecting the text instead of backspacing it). And made these numbers bold to distinguish them.