Simulated annealing optimisation
Tested the simulated annealing with different temperatures and decrease factors to try and figure out the 'best' temperature. The results can be seen in the 'optimising.py' file of this commit in comments at the bottom, along with the process I ended up using to test them. I wanted something which does it in few moves and is consistent, so I figured using the mean and standard deviation of results was perfect.